feat,refactor: disk identity and refactoring
