+++ drive_factory_linux.go
@@ -3,6 +3,8 @@
 
 package go_dparm
 
+import "github.com/jc-lab/go-dparm/common"
+
 type LinuxDriveFactory struct {
 }
 
@@ -14,4 +16,19 @@ func NewLinuxDriveFactory() *LinuxDriveFactory {
 
 func NewSystemDriveFactory() DriveFactory {
 	return NewLinuxDriveFactory()
+}
+
+func (l LinuxDriveFactory) OpenByPath(path string) (DriveHandle, error) {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (l LinuxDriveFactory) EnumDrives() ([]common.DriveInfo, error) {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (l LinuxDriveFactory) EnumVolumes() (EnumVolumeContext, error) {
+	//TODO implement me
+	panic("implement me")
 }
